fix pr feedbacks
