+++ addon/services/base-uploader.ts
@@ -4,71 +4,9 @@ import { isEmpty } from '@ember/utils';
 import UploaderInterface, {
   UploadRequest,
   FileValidator,
-  ValidationResponse,
-  FileSizeRule,
-  FileTypeRule
+  ValidationResponse
 } from '@upfluence/oss-components/types/uploader';
-import parseFilesize from '@upfluence/oss-components/utils/filesize-parser';
-
-interface Validator {
-  validate(request: UploadRequest): { rule: FileValidator; passes: boolean };
-}
-
-class FileSizeValidator implements Validator {
-  static key: string = 'filesize';
-
-  rule: FileSizeRule;
-
-  constructor(rule: FileSizeRule) {
-    this.rule = rule;
-  }
-
-  validate(request: UploadRequest): { rule: FileValidator; passes: boolean } {
-    return {
-      rule: this.rule,
-      passes: parseFilesize(this.rule.value) >= request.file.size
-    };
-  }
-}
-
-class FileTypeValidator implements Validator {
-  static key: string = 'filetype';
-
-  private filetypeTemplates: { [key: string]: string[] } = {
-    image: ['png', 'jpg', 'jpeg']
-  };
-
-  rule: FileTypeRule;
-
-  constructor(rule: FileTypeRule) {
-    this.rule = rule;
-  }
-
-  validate(request: UploadRequest): { rule: FileValidator; passes: boolean } {
-    const filename = request.file.name;
-
-    return {
-      rule: {
-        ...this.rule,
-        value: this.deconstructedFiletypes
-      },
-      passes: this.deconstructedFiletypes.includes(filename.split('.')[filename.split('.').length - 1])
-    };
-  }
-
-  private get deconstructedFiletypes(): string[] {
-    return this.rule.value.reduce((acc: string[], filetype: string) => {
-      let filetypes = [filetype];
-
-      if (Object.keys(this.filetypeTemplates).includes(filetype)) {
-        filetypes = this.filetypeTemplates[filetype];
-      }
-
-      acc = acc.concat(filetypes);
-      return acc;
-    }, []);
-  }
-}
+import { FileSizeValidator, FileTypeValidator, Validator } from '@upfluence/oss-components/utils/uploader-validators';
 
 const AVAILABLE_VALIDATORS = [FileSizeValidator, FileTypeValidator];
